More function

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int n)
 {
-	int len, rem, tmp, i, exp;
+	long int len, rem, tmp, i, exp;
 
 	rem = n;
 	exp = len = 1;
